Readded sunlight

+++ src/geometry.h
@@ -6,6 +6,6 @@ namespace Geometry
 {
     void makeCube(std::vector<float> &vertices, float x, float y, float z, bool faces[6], int type, float light[6][4], float sunlight[6][4]);
     void makeSelectCube(std::vector<float> &vertices, float size);
-    void makePlant(std::vector<float> &vertices, float x, float y, float z, int type, int light, int sunlight);
+    void makePlant(std::vector<float> &vertices, float x, float y, float z, int type, float light, float sunlight);
     void makeGuiQuad(std::vector<float> &vertices, float xs, float ys);
 }
